refactor(poll): move poll mechanism to poll service
